Add ANKI_PREPEND_HEADING property, update toggling
It is retrieved with inheritance and if set takes precedence
over the variable `anki-editor-prepend-heading`.

Also add a function to toggle it and change the behavior
of `anki-editor-toggle-format` to match it: Now every call
to the toggling functions will actually toggle the value in
effect and an entry property is only set if necessary.