Refactor push-notes to use new enqueue/dispatch system.
The previous enqueue/multi system was built on two macros that
were tangled together. Since they had to be nested inside each other
to work we couldn't queue requests across forms. This meant, for
instance, that each function interacting with anki needed at least
one call to the api.

By separating the two concerns of enqueue/dispatch into functions
that only communicate over a request queue we can batch up dozens
or hundreds of individual requests per api call, then dispatch them
at will later. This drastically speeds up operations that create or
update lots of cards.

The cost we pay is that we now have to deal with callbacks, since
we're not processing each note fully all at once. We do the first
half of the work immediately, determing if we need to send the note
to anki and, if so, batching it up, but the response from anki
doesn't come back until we've done this for every other note in the
batch. For this reason we also can't rely on point being at the
relevant note during these callbacks either.

To solve this first issue we can use closures over the note
struct for our callbacks. adding a marker field to the struct
definition solves the second. In this way when processing
success or error we have the note data, its buffer and
position within that buffer, and the response from anki.